split quad, cylinder performance tests
